refactor: 디렉토리명 변경
-Styles to styles
